+++ omfpandas/__init__.py
@@ -7,15 +7,3 @@ try:
 except metadata.PackageNotFoundError:
     # Package is not installed
     pass
-
-try:
-    omf_version = metadata.version('omf')
-    if omf_version.startswith('1'):
-        __omf_version__ = 'v1'
-    elif omf_version.startswith('2'):
-        __omf_version__ = 'v2'
-    else:
-        raise ValueError(f"Unsupported omf version: {omf_version}")
-except metadata.PackageNotFoundError:
-    # OMF package is not installed
-    __omf_version__ = None